+++ GTBack/Services/Restourant/IClientService.cs
@@ -7,7 +7,7 @@ namespace GTBack.Core.Services.Restourant;
 public interface IClientService
 {
     Task<IDataResults<UserDTO>> Me();
-    Task<IDataResults<UserDTO>> GetById(int id);
+    Task<IDataResults<UserDTO>> GetById(long id);
     Task<IDataResults<AuthenticatedUserResponseDto>> Login(LoginDto loginDto);
     Task<IDataResults<AuthenticatedUserResponseDto>> Register(ClientRegisterRequestDTO registerDto);
     Task<IDataResults<AuthenticatedUserResponseDto>> GoogleLogin(GoogleLoginDTO model);
