+++ GTBack/Services/Restourant/IDepartmentService.cs
@@ -6,6 +6,6 @@ namespace GTBack.Core.Services.Restourant;
 
 public interface IDepartmentService<T1,T2>:IGenericService<DepartmentAddDTO,DepartmentListDTO>
 {
-    Task<IDataResults<ICollection<T2>>> ListByCompanyId(int companyId);
+    Task<IDataResults<ICollection<T2>>> ListByCompanyId(long companyId);
 
 }
