+++ GTBack/Services/Restourant/IMenuAndCategoryService.cs
@@ -9,13 +9,13 @@ public interface IMenuAndCategoryService
     Task<IResults> CategoryAdd(CategoryAddOrUpdateDTO model);
     Task<IResults> MenuItemAdd(MenuItemAddOrUpdateDTO model);
     Task<IResults> ExtraMenuItemAdd(ExtraMenuItemAddOrUpdateDTO model);
-    Task<IResults> MenuDelete(int id);
-    Task<IResults> CategoryDelete(int id);
-    Task<IResults> MenuItemDelete(int id);
-    Task<IResults> ExtraMenuItemDelete(int id);
+    Task<IResults> MenuDelete(long id);
+    Task<IResults> CategoryDelete(long id);
+    Task<IResults> MenuItemDelete(long id);
+    Task<IResults> ExtraMenuItemDelete(long id);
     Task<IDataResults<ICollection<CategoryListDTO>>> AllCategoryListByCompanyId();
-    Task<IDataResults<ICollection<MenuItemListDTO>>> MenuItemListByCategoryId(int categoryId);
-    Task<IDataResults<ICollection<ExtraMenuItemListDTO>>> ExtraMenuItemByMenuItemId(int menuItemId);
+    Task<IDataResults<ICollection<MenuItemListDTO>>> MenuItemListByCategoryId(long categoryId);
+    Task<IDataResults<ICollection<ExtraMenuItemListDTO>>> ExtraMenuItemByMenuItemId(long menuItemId);
     Task<IDataResults<ICollection<MenuItemListDTO>>> AllMenuItemsByCompanyId();
 
 }
